refactor: dataBinding 제거
